add local dependency for bookit and update package.json
