+++ src/app/ticket/[id]/page.tsx
@@ -4,11 +4,12 @@ import React, { useState, useEffect } from "react";
 import Image from "next/image";
 import Link from "next/link";
 import { useParams } from "next/navigation";
+import Navbar from "@/app/components/Navbar";
+import Footer from "@/app/components/Footer";
 
 export default function EventDetailPage() {
   const params = useParams();
   const eventId = params.id;
-  const [isScrolled, setIsScrolled] = useState(false);
   const [activeTab, setActiveTab] = useState("description");
   const [ticketQuantity, setTicketQuantity] = useState(1);
   const [selectedTicketType, setSelectedTicketType] = useState("regular");
@@ -95,23 +96,6 @@ export default function EventDetailPage() {
     ],
   };
 
-  // Handle scroll events for sticky header
-  useEffect(() => {
-    const handleScroll = () => {
-      const offset = window.scrollY;
-      if (offset > 100) {
-        setIsScrolled(true);
-      } else {
-        setIsScrolled(false);
-      }
-    };
-
-    window.addEventListener("scroll", handleScroll);
-    return () => {
-      window.removeEventListener("scroll", handleScroll);
-    };
-  }, []);
-
   // Calculate ticket price based on selected ticket type and quantity
   const selectedTicket = event.ticketTypes.find(
     (t) => t.id === selectedTicketType
@@ -126,124 +110,7 @@ export default function EventDetailPage() {
 
   return (
     <div className="min-h-screen bg-gray-50">
-      {/* Sticky Navbar */}
-      <div
-        className={`fixed top-0 left-0 right-0 z-50 transition-all duration-300 ${
-          isScrolled
-            ? "bg-white text-gray-800 shadow-lg"
-            : "bg-gradient-to-r from-indigo-600 via-purple-600 to-pink-500 text-white"
-        }`}
-      >
-        <div className="container mx-auto px-4 py-4">
-          <div className="flex justify-between items-center">
-            <div className="flex items-center space-x-6">
-              <Link href="/" className="font-bold text-2xl flex items-center">
-                <svg
-                  className={`w-8 h-8 mr-2 ${
-                    isScrolled ? "text-indigo-600" : "text-white"
-                  }`}
-                  viewBox="0 0 24 24"
-                  fill="none"
-                  xmlns="http://www.w3.org/2000/svg"
-                >
-                  <path
-                    d="M19 3H5C3.89543 3 3 3.89543 3 5V19C3 20.1046 3.89543 21 5 21H19C20.1046 21 21 20.1046 21 19V5C21 3.89543 20.1046 3 19 3Z"
-                    stroke="currentColor"
-                    strokeWidth="2"
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                  />
-                  <path
-                    d="M9 13L12 16L15 13"
-                    stroke="currentColor"
-                    strokeWidth="2"
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                  />
-                  <path
-                    d="M12 8V16"
-                    stroke="currentColor"
-                    strokeWidth="2"
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                  />
-                </svg>
-                <span
-                  className={
-                    isScrolled ? "text-indigo-600" : "text-white tracking-wider"
-                  }
-                >
-                  BookIt
-                </span>
-              </Link>
-              <nav className="hidden md:flex items-center space-x-1">
-                {[
-                  { name: "Events", href: "/ticket", active: true },
-                  { name: "Create Event", href: "#", active: false },
-                  { name: "Locations", href: "#", active: false },
-                  { name: "Pricing", href: "#", active: false },
-                ].map((item) => (
-                  <Link
-                    key={item.name}
-                    href={item.href}
-                    className={`px-4 py-2 rounded-full text-sm font-medium transition-all duration-200 ${
-                      isScrolled
-                        ? item.active
-                          ? "bg-indigo-100 text-indigo-600"
-                          : "text-gray-700 hover:bg-gray-100"
-                        : item.active
-                        ? "bg-white/20 text-white backdrop-blur-sm"
-                        : "text-white/80 hover:text-white hover:bg-white/10"
-                    }`}
-                  >
-                    {item.name}
-                  </Link>
-                ))}
-              </nav>
-            </div>
-
-            <div className="hidden md:flex items-center space-x-3">
-              <button
-                className={`text-sm py-2 px-5 rounded-full transition-all duration-200 ${
-                  isScrolled
-                    ? "border border-indigo-600 text-indigo-600 hover:bg-indigo-50"
-                    : "text-white border border-white/60 hover:bg-white hover:text-indigo-600"
-                }`}
-              >
-                Masuk
-              </button>
-              <button
-                className={`text-sm py-2 px-5 rounded-full transition-all duration-200 shadow-lg ${
-                  isScrolled
-                    ? "bg-indigo-600 text-white hover:bg-indigo-700"
-                    : "bg-white text-indigo-600 hover:bg-indigo-100"
-                }`}
-              >
-                Daftar
-              </button>
-            </div>
-
-            {/* Mobile menu button */}
-            <button className="md:hidden z-50">
-              <svg
-                xmlns="http://www.w3.org/2000/svg"
-                fill="none"
-                viewBox="0 0 24 24"
-                stroke={isScrolled ? "currentColor" : "white"}
-                className="w-6 h-6"
-              >
-                <path
-                  strokeLinecap="round"
-                  strokeLinejoin="round"
-                  strokeWidth={2}
-                  d="M4 6h16M4 12h16M4 18h16"
-                />
-              </svg>
-            </button>
-          </div>
-        </div>
-      </div>
-
+      <Navbar />
       {/* Event Header with Image */}
       <div className="w-full h-[50vh] relative pt-16">
         <Image
@@ -327,10 +194,14 @@ export default function EventDetailPage() {
               {/* Description Tab */}
               {activeTab === "description" && (
                 <div>
-                  <h2 className="text-2xl font-bold mb-4 text-gray-800">About This Event</h2>
+                  <h2 className="text-2xl font-bold mb-4 text-gray-800">
+                    About This Event
+                  </h2>
                   <p className="text-gray-700 mb-6">{event.description}</p>
 
-                  <h3 className="text-xl font-bold mb-3 text-gray-800">What to Expect</h3>
+                  <h3 className="text-xl font-bold mb-3 text-gray-800">
+                    What to Expect
+                  </h3>
                   <ul className="space-y-2 mb-6">
                     {event.features.map((feature, index) => (
                       <li key={index} className="flex items-start">
@@ -352,7 +223,9 @@ export default function EventDetailPage() {
                     ))}
                   </ul>
 
-                  <h3 className="text-xl font-bold mb-3 text-gray-800">Event Gallery</h3>
+                  <h3 className="text-xl font-bold mb-3 text-gray-800">
+                    Event Gallery
+                  </h3>
                   <div className="grid grid-cols-1 md:grid-cols-3 gap-4 mb-6">
                     {event.gallery.map((image, index) => (
                       <div
@@ -374,7 +247,9 @@ export default function EventDetailPage() {
               {/* Lineup Tab */}
               {activeTab === "lineup" && (
                 <div>
-                  <h2 className="text-2xl font-bold mb-6 text-gray-800">Event Lineup</h2>
+                  <h2 className="text-2xl font-bold mb-6 text-gray-800">
+                    Event Lineup
+                  </h2>
                   <div className="space-y-4">
                     {event.lineup.map((artist, index) => (
                       <div
@@ -382,7 +257,9 @@ export default function EventDetailPage() {
                         className="flex flex-col md:flex-row justify-between p-4 border border-gray-200 rounded-lg hover:border-indigo-200 hover:bg-indigo-50 transition-colors"
                       >
                         <div>
-                          <h3 className="font-bold text-lg text-gray-800">{artist.name}</h3>
+                          <h3 className="font-bold text-lg text-gray-800">
+                            {artist.name}
+                          </h3>
                           <p className="text-gray-600">{artist.day}</p>
                         </div>
                         <div className="flex items-center text-indigo-600 font-medium mt-2 md:mt-0">
@@ -410,10 +287,16 @@ export default function EventDetailPage() {
               {/* Location Tab */}
               {activeTab === "location" && (
                 <div>
-                  <h2 className="text-2xl font-bold mb-4 text-gray-800">Event Location</h2>
+                  <h2 className="text-2xl font-bold mb-4 text-gray-800">
+                    Event Location
+                  </h2>
                   <div className="bg-gray-100 p-4 rounded-lg mb-6">
-                    <h3 className="font-bold mb-2 text-gray-800">{event.location}</h3>
-                    <p className="text-gray-700 mb-4 text-gray-800">{event.address}</p>
+                    <h3 className="font-bold mb-2 text-gray-800">
+                      {event.location}
+                    </h3>
+                    <p className="text-gray-700 mb-4 text-gray-800">
+                      {event.address}
+                    </p>
                     <div className="relative w-full h-64 rounded-lg overflow-hidden">
                       <Image
                         src="/api/placeholder/800/400"
@@ -429,7 +312,9 @@ export default function EventDetailPage() {
                     </div>
                   </div>
 
-                  <h3 className="text-xl font-bold mb-3 text-gray-800">Getting There</h3>
+                  <h3 className="text-xl font-bold mb-3 text-gray-800">
+                    Getting There
+                  </h3>
                   <div className="space-y-4">
                     <div className="flex items-start">
                       <div className="flex-shrink-0 w-10 h-10 bg-indigo-100 rounded-full flex items-center justify-center mr-4">
@@ -477,7 +362,9 @@ export default function EventDetailPage() {
                         </svg>
                       </div>
                       <div>
-                        <h4 className="font-bold mb-1 text-gray-800">Parking</h4>
+                        <h4 className="font-bold mb-1 text-gray-800">
+                          Parking
+                        </h4>
                         <p className="text-gray-700">
                           On-site parking available for Rp 25.000 per entry
                         </p>
@@ -491,7 +378,9 @@ export default function EventDetailPage() {
               {activeTab === "reviews" && (
                 <div>
                   <div className="flex justify-between items-center mb-6">
-                    <h2 className="text-2xl font-bold text-gray-800">Reviews & Ratings</h2>
+                    <h2 className="text-2xl font-bold text-gray-800">
+                      Reviews & Ratings
+                    </h2>
                     <button className="text-indigo-600 font-medium hover:text-indigo-800 transition-colors">
                       Write a Review
                     </button>
@@ -537,7 +426,9 @@ export default function EventDetailPage() {
           {/* Right Column - Ticket Selection */}
           <div className="lg:w-1/3 mt-8 lg:mt-0">
             <div className="bg-white p-6 rounded-xl shadow-sm sticky top-24">
-              <h2 className="text-xl font-bold mb-4 text-gray-800">Select Tickets</h2>
+              <h2 className="text-xl font-bold mb-4 text-gray-800">
+                Select Tickets
+              </h2>
 
               {/* Ticket Type Selection */}
               <div className="space-y-4 mb-6">
@@ -777,7 +668,9 @@ export default function EventDetailPage() {
       {/* Related Events */}
       <div className="bg-gray-50 py-12">
         <div className="container mx-auto px-4">
-          <h2 className="text-2xl font-bold mb-8 text-gray-800">You Might Also Like</h2>
+          <h2 className="text-2xl font-bold mb-8 text-gray-800">
+            You Might Also Like
+          </h2>
           <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-4 gap-6">
             {[1, 2, 3, 4].map((item) => (
               <div
@@ -855,289 +748,7 @@ export default function EventDetailPage() {
         </div>
       </div>
 
-      {/* Footer */}
-      <footer className="bg-gray-900 text-white py-16">
-        <div className="container mx-auto px-4">
-          <div className="grid grid-cols-1 md:grid-cols-4 gap-10">
-            <div>
-              <div className="flex items-center mb-6">
-                <svg
-                  className="w-8 h-8 mr-2"
-                  viewBox="0 0 24 24"
-                  fill="none"
-                  xmlns="http://www.w3.org/2000/svg"
-                >
-                  <path
-                    d="M19 3H5C3.89543 3 3 3.89543 3 5V19C3 20.1046 3.89543 21 5 21H19C20.1046 21 21 20.1046 21 19V5C21 3.89543 20.1046 3 19 3Z"
-                    stroke="currentColor"
-                    strokeWidth="2"
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                  />
-                  <path
-                    d="M9 13L12 16L15 13"
-                    stroke="currentColor"
-                    strokeWidth="2"
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                  />
-                  <path
-                    d="M12 8V16"
-                    stroke="currentColor"
-                    strokeWidth="2"
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                  />
-                </svg>
-                <span className="text-xl font-bold">BookIt</span>
-              </div>
-              <p className="text-gray-400 text-sm mb-6">
-                Platform pembelian tiket online terpercaya untuk berbagai event
-                di seluruh Indonesia.
-              </p>
-              <div className="flex space-x-4">
-                <a
-                  href="#"
-                  className="bg-gray-800 w-10 h-10 rounded-full flex items-center justify-center hover:bg-indigo-600 transition-all duration-300"
-                >
-                  <svg
-                    fill="currentColor"
-                    viewBox="0 0 24 24"
-                    className="w-5 h-5"
-                  >
-                    <path d="M9 8h-3v4h3v12h5v-12h3.642l.358-4h-4v-1.667c0-.955.192-1.333 1.115-1.333h2.885v-5h-3.808c-3.596 0-5.192 1.583-5.192 4.615v3.385z" />
-                  </svg>
-                </a>
-                <a
-                  href="#"
-                  className="bg-gray-800 w-10 h-10 rounded-full flex items-center justify-center hover:bg-blue-500 transition-all duration-300"
-                >
-                  <svg
-                    fill="currentColor"
-                    viewBox="0 0 24 24"
-                    className="w-5 h-5"
-                  >
-                    <path d="M24 4.557c-.883.392-1.832.656-2.828.775 1.017-.609 1.798-1.574 2.165-2.724-.951.564-2.005.974-3.127 1.195-.897-.957-2.178-1.555-3.594-1.555-3.179 0-5.515 2.966-4.797 6.045-4.091-.205-7.719-2.165-10.148-5.144-1.29 2.213-.669 5.108 1.523 6.574-.806-.026-1.566-.247-2.229-.616-.054 2.281 1.581 4.415 3.949 4.89-.693.188-1.452.232-2.224.084.626 1.956 2.444 3.379 4.6 3.419-2.07 1.623-4.678 2.348-7.29 2.04 2.179 1.397 4.768 2.212 7.548 2.212 9.142 0 14.307-7.721 13.995-14.646.962-.695 1.797-1.562 2.457-2.549z" />
-                  </svg>
-                </a>
-                <a
-                  href="#"
-                  className="bg-gray-800 w-10 h-10 rounded-full flex items-center justify-center hover:bg-pink-600 transition-all duration-300"
-                >
-                  <svg
-                    fill="currentColor"
-                    viewBox="0 0 24 24"
-                    className="w-5 h-5"
-                  >
-                    <path d="M12 2.163c3.204 0 3.584.012 4.85.07 3.252.148 4.771 1.691 4.919 4.919.058 1.265.069 1.645.069 4.849 0 3.205-.012 3.584-.069 4.849-.149 3.225-1.664 4.771-4.919 4.919-1.266.058-1.644.07-4.85.07-3.204 0-3.584-.012-4.849-.07-3.26-.149-4.771-1.699-4.919-4.92-.058-1.265-.07-1.644-.07-4.849 0-3.204.013-3.583.07-4.849.149-3.227 1.664-4.771 4.919-4.919 1.266-.057 1.645-.069 4.849-.069zm0-2.163c-3.259 0-3.667.014-4.947.072-4.358.2-6.78 2.618-6.98 6.98-.059 1.281-.073 1.689-.073 4.948 0 3.259.014 3.668.072 4.948.2 4.358 2.618 6.78 6.98 6.98 1.281.058 1.689.072 4.948.072 3.259 0 3.668-.014 4.948-.072 4.354-.2 6.782-2.618 6.979-6.98.059-1.28.073-1.689.073-4.948 0-3.259-.014-3.667-.072-4.947-.196-4.354-2.617-6.78-6.979-6.98-1.281-.059-1.69-.073-4.949-.073zm0 5.838c-3.403 0-6.162 2.759-6.162 6.162s2.759 6.163 6.162 6.163 6.162-2.759 6.162-6.163c0-3.403-2.759-6.162-6.162-6.162zm0 10.162c-2.209 0-4-1.79-4-4 0-2.209 1.791-4 4-4s4 1.791 4 4c0 2.21-1.791 4-4 4zm6.406-11.845c-.796 0-1.441.645-1.441 1.44s.645 1.44 1.441 1.44c.795 0 1.439-.645 1.439-1.44s-.644-1.44-1.439-1.44z" />
-                  </svg>
-                </a>
-              </div>
-            </div>
-
-            <div>
-              <h4 className="font-bold text-lg mb-4">Bantuan</h4>
-              <ul className="space-y-3 text-sm text-gray-400">
-                <li>
-                  <a
-                    href="#"
-                    className="hover:text-indigo-400 transition-colors duration-200 flex items-center"
-                  >
-                    <svg
-                      className="w-4 h-4 mr-2"
-                      fill="none"
-                      stroke="currentColor"
-                      viewBox="0 0 24 24"
-                    >
-                      <path
-                        strokeLinecap="round"
-                        strokeLinejoin="round"
-                        strokeWidth="2"
-                        d="M9 5l7 7-7 7"
-                      ></path>
-                    </svg>
-                    Cara Pembelian
-                  </a>
-                </li>
-                <li>
-                  <a
-                    href="#"
-                    className="hover:text-indigo-400 transition-colors duration-200 flex items-center"
-                  >
-                    <svg
-                      className="w-4 h-4 mr-2"
-                      fill="none"
-                      stroke="currentColor"
-                      viewBox="0 0 24 24"
-                    >
-                      <path
-                        strokeLinecap="round"
-                        strokeLinejoin="round"
-                        strokeWidth="2"
-                        d="M9 5l7 7-7 7"
-                      ></path>
-                    </svg>
-                    Cara Pembayaran
-                  </a>
-                </li>
-                <li>
-                  <a
-                    href="#"
-                    className="hover:text-indigo-400 transition-colors duration-200 flex items-center"
-                  >
-                    <svg
-                      className="w-4 h-4 mr-2"
-                      fill="none"
-                      stroke="currentColor"
-                      viewBox="0 0 24 24"
-                    >
-                      <path
-                        strokeLinecap="round"
-                        strokeLinejoin="round"
-                        strokeWidth="2"
-                        d="M9 5l7 7-7 7"
-                      ></path>
-                    </svg>
-                    FAQ
-                  </a>
-                </li>
-                <li>
-                  <a
-                    href="#"
-                    className="hover:text-indigo-400 transition-colors duration-200 flex items-center"
-                  >
-                    <svg
-                      className="w-4 h-4 mr-2"
-                      fill="none"
-                      stroke="currentColor"
-                      viewBox="0 0 24 24"
-                    >
-                      <path
-                        strokeLinecap="round"
-                        strokeLinejoin="round"
-                        strokeWidth="2"
-                        d="M9 5l7 7-7 7"
-                      ></path>
-                    </svg>
-                    Hubungi Kami
-                  </a>
-                </li>
-              </ul>
-            </div>
-
-            <div>
-              <h4 className="font-bold text-lg mb-4">Tentang</h4>
-              <ul className="space-y-3 text-sm text-gray-400">
-                <li>
-                  <a
-                    href="#"
-                    className="hover:text-indigo-400 transition-colors duration-200 flex items-center"
-                  >
-                    <svg
-                      className="w-4 h-4 mr-2"
-                      fill="none"
-                      stroke="currentColor"
-                      viewBox="0 0 24 24"
-                    >
-                      <path
-                        strokeLinecap="round"
-                        strokeLinejoin="round"
-                        strokeWidth="2"
-                        d="M9 5l7 7-7 7"
-                      ></path>
-                    </svg>
-                    Tentang Kami
-                  </a>
-                </li>
-                <li>
-                  <a
-                    href="#"
-                    className="hover:text-indigo-400 transition-colors duration-200 flex items-center"
-                  >
-                    <svg
-                      className="w-4 h-4 mr-2"
-                      fill="none"
-                      stroke="currentColor"
-                      viewBox="0 0 24 24"
-                    >
-                      <path
-                        strokeLinecap="round"
-                        strokeLinejoin="round"
-                        strokeWidth="2"
-                        d="M9 5l7 7-7 7"
-                      ></path>
-                    </svg>
-                    Karir
-                  </a>
-                </li>
-                <li>
-                  <a
-                    href="#"
-                    className="hover:text-indigo-400 transition-colors duration-200 flex items-center"
-                  >
-                    <svg
-                      className="w-4 h-4 mr-2"
-                      fill="none"
-                      stroke="currentColor"
-                      viewBox="0 0 24 24"
-                    >
-                      <path
-                        strokeLinecap="round"
-                        strokeLinejoin="round"
-                        strokeWidth="2"
-                        d="M9 5l7 7-7 7"
-                      ></path>
-                    </svg>
-                    Kebijakan Privasi
-                  </a>
-                </li>
-                <li>
-                  <a
-                    href="#"
-                    className="hover:text-indigo-400 transition-colors duration-200 flex items-center"
-                  >
-                    <svg
-                      className="w-4 h-4 mr-2"
-                      fill="none"
-                      stroke="currentColor"
-                      viewBox="0 0 24 24"
-                    >
-                      <path
-                        strokeLinecap="round"
-                        strokeLinejoin="round"
-                        strokeWidth="2"
-                        d="M9 5l7 7-7 7"
-                      ></path>
-                    </svg>
-                    Syarat dan Ketentuan
-                  </a>
-                </li>
-              </ul>
-            </div>
-
-            <div>
-              <h4 className="font-bold text-lg mb-4">Ikuti Kami</h4>
-              <p className="text-gray-400 text-sm mb-4">
-                Dapatkan info terbaru tentang event dan promo spesial
-              </p>
-              <div className="flex items-center">
-                <input
-                  type="email"
-                  placeholder="Email Kamu"
-                  className="py-2 px-4 bg-gray-800 rounded-l-lg focus:outline-none focus:ring-2 focus:ring-indigo-500 text-white text-sm flex-grow"
-                />
-                <button className="bg-gradient-to-r from-indigo-600 to-purple-600 text-white py-2 px-4 rounded-r-lg font-medium text-sm">
-                  Subscribe
-                </button>
-              </div>
-            </div>
-          </div>
-
-          <div className="border-t border-gray-800 mt-12 pt-8 text-center text-sm text-gray-400">
-            <p>© 2025 BookIt. All rights reserved.</p>
-          </div>
-        </div>
-      </footer>
+      <Footer />
     </div>
   );
 }
